pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "netlicensing-mcp"
version = "1.0.0"
description = "MCP server for the Labs64 NetLicensing license management API"
readme = "README.md"
license = { text = "Apache-2.0" }
requires-python = ">=3.12"
keywords = ["mcp", "netlicensing", "license-management", "labs64", "fastmcp"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.14",
    "Topic :: Software Development :: Libraries",
]
dependencies = [
    "mcp[cli]>=1.7.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "ruff",
    "mypy",
    "pip-audit",
]

[project.urls]
Homepage = "https://github.com/yourusername/netlicensing-mcp"
Repository = "https://github.com/yourusername/netlicensing-mcp"
Issues = "https://github.com/yourusername/netlicensing-mcp/issues"

[project.scripts]
netlicensing-mcp = "netlicensing_mcp.server:main"

[tool.hatch.build.targets.wheel]
packages = ["src/netlicensing_mcp"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "--cov=src/netlicensing_mcp --cov-report=term-missing --cov-report=xml"

[tool.coverage.run]
source = ["src/netlicensing_mcp"]
omit = ["tests/*"]

[tool.ruff]
line-length = 100
target-version = "py311"
